refactor(server): move the terminal's policy and protocol under platform/
Layer relocation, no behaviour change: the hmr README's four-layer model
puts business APIs and behaviour policy in the platform, and hmr/
resources.ts is explicit that spawning a shell "is the platform's
business and lives there". The terminal arrived entirely in the runtime,
which is why a hot push cannot carry it — the reported 404 was a Web App
newer than the runtime serving it.

This first step is the file move with the wiring untouched (app.ts still
mounts the routes): routes, manager, session, snapshot, frames,
output-coalescer, input-mode and size-ownership now live under
platform/terminal/. ws.ts and spawn-helper.ts stay in the runtime — the
seam cannot carry a stream, and the spawn-helper repair is a machine
primitive.
