+++ packages/server/src/app.ts
@@ -58,8 +58,8 @@ import { agentTracesRoutes } from "./http/routes/agent-traces.js";
 import { usageRoutes } from "./http/routes/usage.js";
 import { agentSessionsRoutes, sessionsRoutes } from "./http/routes/sessions.js";
 import { versionRoutes } from "./http/routes/version.js";
-import { terminalsRoutes } from "./http/routes/terminals.js";
-import { TerminalManager } from "./terminal/manager.js";
+import { terminalsRoutes } from "./platform/terminal/routes.js";
+import { TerminalManager } from "./platform/terminal/manager.js";
 import { ChannelHub } from "./runtime/channel.js";
 import { ErrorRecorder } from "./runtime/error-recorder.js";
 import { createCoreSessionLoader, SessionManager } from "./runtime/session-manager.js";
