+++ packages/server/src/platform/terminal/frames.ts
