+++ packages/server/src/platform/terminal/output-coalescer.ts
