+++ packages/server/src/platform/terminal/size-ownership.ts
