+++ packages/server/src/platform/terminal/snapshot.ts
